+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.32.6](https://github.com/matter-labs/zksync-crypto/compare/v0.32.5...v0.32.6) (2025-10-28)
+
+
+### Features
+
+* Remove panics from verifier ([#115](https://github.com/matter-labs/zksync-crypto/issues/115)) ([400cf6b](https://github.com/matter-labs/zksync-crypto/commit/400cf6b107d21053ab37ef7c746feeaecc0d7aad))
+
+## [0.32.5](https://github.com/matter-labs/zksync-crypto/compare/v0.32.4...v0.32.5) (2025-10-17)
+
+
+### Features
+
+* Allow zk in circuit configuration ([#113](https://github.com/matter-labs/zksync-crypto/issues/113)) ([ac35872](https://github.com/matter-labs/zksync-crypto/commit/ac358729a4ccd5a69827a5094d5667b8353a42c2))
+
+## [0.32.4](https://github.com/matter-labs/zksync-crypto/compare/v0.32.3...v0.32.4) (2025-10-15)
+
+
+### Features
+
+* Allow zk in circuit configuration ([#111](https://github.com/matter-labs/zksync-crypto/issues/111)) ([ded46e1](https://github.com/matter-labs/zksync-crypto/commit/ded46e18b251324e671764e5b631c1b65b8ca9b6))
+
 ## [0.32.3](https://github.com/matter-labs/zksync-crypto/compare/v0.32.2...v0.32.3) (2025-08-22)